Cleanup: split PRTModifierEnum implementation into separate file
